Extract save api call to own function in utils
